Added more util functions AND added DateRange type
